requirements.txt
groq>=0.9.0
streamlit>=1.35.0
pydantic>=2.7.0
pymupdf>=1.24.0
python-dotenv>=1.0.0
loguru>=0.7.2
httpx>=0.27.0
pandas>=2.2.0
matplotlib>=3.8.0
plotly>=5.22.0
reportlab>=4.2.0
